feat: exposed positioned indicator container
