+++ lib/src/widgets/custom_material_indicator.dart
@@ -367,7 +367,7 @@ class _CustomMaterialIndicatorState extends State<CustomMaterialIndicator> {
         return Stack(
           children: <Widget>[
             widget.scrollableBuilder(context, child, controller),
-            _PositionedIndicatorContainer(
+            PositionedIndicatorContainer(
               edgeOffset: widget.edgeOffset,
               displacement: widget.displacement,
               controller: controller,
@@ -390,171 +390,4 @@ class _CustomMaterialIndicatorState extends State<CustomMaterialIndicator> {
     _internalIndicatorController?.dispose();
     super.dispose();
   }
-}
-
-class _PositionedIndicatorContainer extends StatelessWidget {
-  final IndicatorController controller;
-  final double displacement;
-  final Widget child;
-  final double edgeOffset;
-
-  /// Position child widget in a similar way
-  /// to the built-in [RefreshIndicator] widget.
-  const _PositionedIndicatorContainer({
-    required this.child,
-    required this.controller,
-    required this.displacement,
-    required this.edgeOffset,
-  });
-
-  Alignment _getAlignment(IndicatorSide side) {
-    switch (side) {
-      case IndicatorSide.left:
-        return Alignment.centerLeft;
-      case IndicatorSide.top:
-        return Alignment.topCenter;
-      case IndicatorSide.right:
-        return Alignment.centerRight;
-      case IndicatorSide.bottom:
-        return Alignment.bottomCenter;
-      case IndicatorSide.none:
-        throw UnsupportedError('Cannot get alignment for "none" side.');
-    }
-  }
-
-  EdgeInsets _getEdgeInsets(IndicatorSide side) {
-    switch (side) {
-      case IndicatorSide.left:
-        return EdgeInsets.only(left: displacement);
-      case IndicatorSide.top:
-        return EdgeInsets.only(top: displacement);
-      case IndicatorSide.right:
-        return EdgeInsets.only(right: displacement);
-      case IndicatorSide.bottom:
-        return EdgeInsets.only(bottom: displacement);
-      case IndicatorSide.none:
-        throw UnsupportedError('Cannot get edge insets for "none" side.');
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final side = controller.side;
-    if (side.isNone) return const SizedBox.shrink();
-
-    final isVerticalAxis = side.isTop || side.isBottom;
-    final isHorizontalAxis = side.isLeft || side.isRight;
-
-    final AlignmentDirectional alignment = isVerticalAxis
-        ? AlignmentDirectional(-1.0, side.isTop ? 1.0 : -1.0)
-        : AlignmentDirectional(side.isLeft ? 1.0 : -1.0, -1.0);
-
-    final endOffset = isVerticalAxis
-        ? Offset(0.0, side.isTop ? 1.0 : -1.0)
-        : Offset(side.isLeft ? 1.0 : -1.0, 0.0);
-
-    final animation = controller.isFinalizing
-        ? AlwaysStoppedAnimation(endOffset)
-        : Tween(begin: const Offset(0.0, 0.0), end: endOffset)
-            .animate(controller);
-
-    return Positioned(
-      top: isHorizontalAxis
-          ? 0
-          : side.isTop
-              ? edgeOffset
-              : null,
-      bottom: isHorizontalAxis
-          ? 0
-          : side.isBottom
-              ? edgeOffset
-              : null,
-      left: isVerticalAxis
-          ? 0
-          : side.isLeft
-              ? edgeOffset
-              : null,
-      right: isVerticalAxis
-          ? 0
-          : side.isRight
-              ? edgeOffset
-              : null,
-      child: Align(
-        alignment: alignment,
-        heightFactor: isVerticalAxis ? 0.0 : null,
-        widthFactor: isHorizontalAxis ? 0.0 : null,
-        child: SlideTransition(
-          position: animation,
-          child: Padding(
-            padding: _getEdgeInsets(side),
-            child: Align(
-              alignment: _getAlignment(side),
-              child: child,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _InfiniteRotation extends StatefulWidget {
-  final Widget? child;
-  final bool running;
-
-  const _InfiniteRotation({
-    required this.child,
-    required this.running,
-  });
-
-  @override
-  _InfiniteRotationState createState() => _InfiniteRotationState();
-}
-
-class _InfiniteRotationState extends State<_InfiniteRotation>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _rotationController;
-
-  @override
-  void didUpdateWidget(_InfiniteRotation oldWidget) {
-    if (oldWidget.running != widget.running) {
-      if (widget.running) {
-        _startAnimation();
-      } else {
-        _rotationController.animateTo(
-          0.0,
-          duration: const Duration(milliseconds: 50),
-        );
-      }
-    }
-    super.didUpdateWidget(oldWidget);
-  }
-
-  @override
-  void initState() {
-    _rotationController = AnimationController(
-      duration: const Duration(milliseconds: 750),
-      vsync: this,
-    );
-
-    if (widget.running) {
-      _startAnimation();
-    }
-
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    _rotationController.dispose();
-    super.dispose();
-  }
-
-  void _startAnimation() {
-    _rotationController.repeat();
-  }
-
-  @override
-  Widget build(BuildContext context) =>
-      RotationTransition(turns: _rotationController, child: widget.child);
 }
