+++ lib/src/widgets/widgets.dart
@@ -1,1 +1,2 @@
 export 'custom_material_indicator.dart';
+export 'positioned_indicator_container.dart';
